Development

+++ src/main/java/com/fnrsportfolio/negraeportfolio/repository/EducacionRepository.java
@@ -1,15 +1,14 @@
 package com.fnrsportfolio.negraeportfolio.repository;
 
 import com.fnrsportfolio.negraeportfolio.model.Educacion;
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EducacionRepository extends JpaRepository<Educacion, Integer> {
+public interface EducacionRepository extends JpaRepository<Educacion, Long> {
 /*
     public Optional<Educacion> findByNombreEdu(String nombreEdu);
 
     public boolean existByNombreEdu(String nombreEdu);
-    */
+*/
 }
